% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_style_widget}
\alias{generate_style_widget}
\title{Generate documentation for style widgets}
\usage{
generate_style_widget(
  name = "ButtonStyle",
  formals = c(),
  error_call = current_env()
)
}
\arguments{
\item{name}{name of the model}

\item{formals}{preferred formals order}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Generate documentation for style widgets
}
