% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jupyterwidgetmodels.R
\docType{data}
\name{jupyterwidgetmodels}
\alias{jupyterwidgetmodels}
\title{Jupyter Widget Models}
\format{
\subsection{\code{jupyterwidgetmodels}}{

A data frame with 69 and 8 columns
\describe{
\item{_view_module}{Name of the view module}
\item{_view_module_version}{version of the view module}
\item{_view_name}{Name of the view}
\item{_model_module}{Name of the model module}
\item{_model_module_version}{version of the model module}
\item{_model_name}{Name of the model}
\item{dom}{TRUE if this is a DOM widget}
\item{attributes}{List of attributes}
...
}
}
}
\source{
\url{https://github.com/jupyter-widgets/ipywidgets/blob/main/packages/schema/jupyterwidgetmodels.latest.json}
}
\usage{
jupyterwidgetmodels
}
\description{
Jupyter Widget Models
}
\examples{
# attributes for Button
jupyterwidgetmodels$attributes[[7]]

}
\keyword{datasets}
